backend/pdf_extractor/requirements.txt
PyPDF2==3.0.1
requests==2.28.2
python-dotenv==1.0.0
openai==1.6.1
tqdm==4.66.1